+++ NEWS.md
@@ -1,5 +1,9 @@
 # Current
 
+* new: listToDepth() 
+* new small helper functions: backspace(), left(), orderBy(), repeated(), 
+  right(), space()
+
 # [kwb.utils 0.5.1](https://github.com/KWB-R/kwb.utils/releases/tag/v0.5.1) <small>2020-04-20</small>
 
 * Fix bug in moveColumnsToFront(): keep data frame structure in case that the